requirements.txt
asgiref==3.2.7
dj-database-url==0.5.0
Django==3.0.7
django-heroku==0.3.1
django-icons==1.1.2
django-widget-tweaks==1.4.8
djangorestframework==3.11.0
gunicorn==20.0.4
Pillow==7.1.1
psycopg2==2.8.5
pytz==2019.3
sqlparse==0.3.1
whitenoise==5.1.0
